feat(websocket_server sink): add support for websocket subprotocol (#22854)
* feat(websocket_server sink): add support for websocket subprotocol

Adds support for `Sec-WebSocket-Protocol` header in websocket connection request. By default (and
currently, before this feature), the server ignored this header and never put it in responses. It
can be configured to accept any subprotocol and put it in response header (some clients will require
this) and it can also be configured to only accept some subprotocols and ignore others. This only
affects the connection requests, everything afterwards functions the same.

* Add changelog entry

* Update changelog.d/22854_websocket_server_subprotocol_support.feature.md

---------

Co-authored-by: Pavlos Rontidis <pavlos.rontidis@gmail.com>